+++ NEWS.md
@@ -2,6 +2,8 @@ v.0.1.4
 =======
 * new plots for PLS model: `plotxloadings()`, `plotxyloadings()` and `plotweights()`
 * small improvements for regression coefficients plot
+* methods `scatter()` and `plot()` now have a new parameter `'Groupby'` for easy color grouping
+* method `catter()` has a new parameter `'ShowContour'` for drawing contour for a cluster of points
 
 v.0.1.3
 ========